pyproject.toml
[project]
name = "checkmk_ansible_collection"
version = "6.3.0"
requires-python = ">=3.11"
dependencies = [
    "ansible-core==2.19.4",
    "antsibull-changelog==0.34.0",
    "antsibull-docs==2.22.0",
    "ansible-lint==25.9.2",
    "jinja2==3.1.6",
    "molecule==25.9.0",
    "molecule-plugins[podman]==25.8.12",
    "pywinrm==0.5.0",
    "yamllint==1.37.1",
]

[dependency-groups]
dev = [
    "black==25.11.0",
    "click==8.3.0",
    "isort==7.0.0",
    "mypy-extensions==1.1.0",
    "pathspec==0.12.1",
    "platformdirs==4.5.0",
    "tomli==2.3.0",
    "typing-extensions==4.15.0",
]


[tool.isort]
profile = "black"
